Corringo problemas de associação

+++ src/main/java/com/danielalves/cvu/repositories/VeiculoRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.danielalves.cvu.domain.Veiculo;
 
-public interface VeiculoRepository extends JpaRepository<Veiculo, Integer> {
+public interface VeiculoRepository extends JpaRepository<Veiculo, Long> {
 
 }
