+++ src/main/java/com/danielalves/cvu/resources/UsuarioResource.java
@@ -27,7 +27,7 @@ public class UsuarioResource {
 	private UsuarioService service;
 	
 	@RequestMapping(value="/{id}", method=RequestMethod.GET)
-	public ResponseEntity<Usuario> find(@PathVariable Integer id) {
+	public ResponseEntity<Usuario> find(@PathVariable Long id) {
 		
 		Usuario obj = service.find(id);
 		
@@ -48,7 +48,7 @@ public class UsuarioResource {
 	}
 	
 	@RequestMapping(value="/{id}", method=RequestMethod.PUT)
-	public ResponseEntity<Void> update(@RequestBody Usuario obj, @PathVariable Integer id) {
+	public ResponseEntity<Void> update(@RequestBody Usuario obj, @PathVariable Long id) {
 		
 		obj.setId(id);
 		obj = service.update(obj);
@@ -58,7 +58,7 @@ public class UsuarioResource {
 	}
 	
 	@RequestMapping(value="/{id}", method=RequestMethod.DELETE)
-	public ResponseEntity<Void> delete(@PathVariable Integer id) {
+	public ResponseEntity<Void> delete(@PathVariable Long id) {
 		
 		service.delete(id);
